Optimisation and refactor
