+++ common/src/main/java/net/favouriteless/enchanted/common/blocks/cauldrons/CauldronBlockBase.java
@@ -1,7 +1,7 @@
 package net.favouriteless.enchanted.common.blocks.cauldrons;
 
 import net.favouriteless.enchanted.common.blocks.entity.CauldronBlockEntity;
-import net.favouriteless.enchanted.util.PlayerInventoryHelper;
+import net.favouriteless.enchanted.common.util.PlayerInventoryHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
