+++ common/src/main/java/net/favouriteless/enchanted/common/circle_magic/rites/ProtectionWaystoneRite.java
@@ -1,6 +1,6 @@
 package net.favouriteless.enchanted.common.circle_magic.rites;
 
-import net.favouriteless.enchanted.util.WaystoneHelper;
+import net.favouriteless.enchanted.common.util.WaystoneHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
