+++ common/src/main/java/net/favouriteless/enchanted/common/items/BoundWaystoneItem.java
@@ -1,6 +1,6 @@
 package net.favouriteless.enchanted.common.items;
 
-import net.favouriteless.enchanted.util.WaystoneHelper;
+import net.favouriteless.enchanted.common.util.WaystoneHelper;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
