+++ common/src/main/java/net/favouriteless/enchanted/common/menus/SpinningWheelMenu.java
@@ -5,7 +5,7 @@ import net.favouriteless.enchanted.common.init.registry.EBlocks;
 import net.favouriteless.enchanted.common.init.registry.EMenuTypes;
 import net.favouriteless.enchanted.common.menus.slots.NonJarInputSlot;
 import net.favouriteless.enchanted.common.menus.slots.OutputSlot;
-import net.favouriteless.enchanted.util.MenuUtils;
+import net.favouriteless.enchanted.common.util.MenuUtils;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
