+++ common/src/main/java/net/favouriteless/enchanted/common/menus/WitchOvenMenu.java
@@ -8,8 +8,7 @@ import net.favouriteless.enchanted.common.menus.slots.FuelSlot;
 import net.favouriteless.enchanted.common.menus.slots.JarInputSlot;
 import net.favouriteless.enchanted.common.menus.slots.NonJarInputSlot;
 import net.favouriteless.enchanted.common.menus.slots.OutputSlot;
-import net.favouriteless.enchanted.util.ItemStackHelper;
-import net.favouriteless.enchanted.util.MenuUtils;
+import net.favouriteless.enchanted.common.util.MenuUtils;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.player.Inventory;
