+++ common/src/main/java/net/favouriteless/enchanted/common/menus/slots/FuelSlot.java
@@ -1,6 +1,5 @@
 package net.favouriteless.enchanted.common.menus.slots;
 
-import net.favouriteless.enchanted.util.ItemStackHelper;
 import net.minecraft.world.Container;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
