+++ common/src/main/java/net/favouriteless/enchanted/common/util/ItemUtils.java
@@ -1,14 +1,19 @@
 package net.favouriteless.enchanted.common.util;
 
+import com.google.gson.JsonObject;
 import net.favouriteless.enchanted.common.Enchanted;
 import net.favouriteless.enchanted.platform.CommonServices;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NumericTag;
 import net.minecraft.nbt.Tag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.GsonHelper;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
@@ -59,8 +64,16 @@ public class ItemUtils {
         if(a.getItem() != b.getItem())
             return false;
 
-        CompoundTag aTags = a.copy().getOrCreateTag();
-        CompoundTag bTags = b.copy().getOrCreateTag();
+        if(b.hasTag()) {
+            if(!a.hasTag())
+                return false;
+        }
+        else {
+            return true;
+        }
+
+        CompoundTag aTags = a.getTag();
+        CompoundTag bTags = b.getTag();
         for(String tag : bTags.getAllKeys()) {
             if(!aTags.contains(tag))
                 return false;
@@ -75,13 +88,13 @@ public class ItemUtils {
 
             if(!Objects.equals(aTag, bTag)) {
                 if(aTag instanceof NumericTag aNumeric && bTag instanceof NumericTag bNumeric) {
-                    //if numeric, check equal
+                    // If numeric, check equal
                     if(aNumeric.getAsFloat() == bNumeric.getAsFloat())
                         continue;
                 }
 
                 if(aTag instanceof CompoundTag aCompound && bTag instanceof CompoundTag bCompound) {
-                    //if compound, convert snbt to json, then check equal
+                    // If compound, convert SNBT to json, then check equal
                     if(snbtToJson(aCompound.toString()).equals(snbtToJson(bCompound.toString())))
                         continue;
                 }
@@ -105,4 +118,58 @@ public class ItemUtils {
     public static String snbtToJson(String snbt) {
         return snbt.replaceAll("([0-9])[bBsSlLfFdD]([,}])", "$1$2");
     }
+
+    /**
+     * Parse a given {@link JsonObject} to get an {@link ItemStack}.
+     *
+     * @param json The {@link JsonObject} to parse.
+     * @param readNbt True if NBT data should be parsed.
+     * @return The {@link ItemStack} which was stored in json.
+     */
+    public static ItemStack fromJson(JsonObject json, boolean readNbt) {
+        Item item = BuiltInRegistries.ITEM.get(new ResourceLocation(GsonHelper.getAsString(json, "item")));
+        int count = GsonHelper.getAsInt(json, "count", 1); // Get count, default to 1 if there was no count.
+        ItemStack out = new ItemStack(item, count);
+
+        if (readNbt && json.has("nbt"))
+            out.setTag(JsonHelper.readTag(json.get("nbt")));
+
+        return out;
+    }
+
+    /**
+     * Serialize a given {@link ItemStack} to a {@link JsonObject}.
+     *
+     * @param item The {@link ItemStack} to serialize.
+     * @param saveNbt True if item's {@link CompoundTag} should be serialized.
+     * @return A {@link JsonObject} containing item.
+     */
+    public static JsonObject asJson(ItemStack item, boolean saveNbt) {
+        return asJson(item.getItem(), item.getCount(), saveNbt && item.hasTag() ? item.getTag() : null);
+    }
+
+    /**
+     * Serialize a given {@link Item}, count and {@link CompoundTag} to a {@link JsonObject} in the same format as an
+     * {@link ItemStack}.
+     *
+     * @param item The {@link Item} of the {@link ItemStack}.
+     * @param count The count of the {@link ItemStack}.
+     * @return A {@link JsonObject} containing item.
+     */
+    public static JsonObject asJson(Item item, int count, CompoundTag nbt) {
+        JsonObject obj = new JsonObject();
+
+        obj.addProperty("item", BuiltInRegistries.ITEM.getKey(item).toString());
+        if(count > 1)
+            obj.addProperty("count", count);
+
+        if(nbt != null) {
+            CompoundTag noDamage = nbt.copy();
+            noDamage.remove("Damage");
+            if(!noDamage.isEmpty())
+                obj.addProperty("nbt", noDamage.getAsString());
+        }
+        return obj;
+    }
+
 }
