+++ common/src/main/java/net/favouriteless/enchanted/common/util/JsonHelper.java
@@ -1,4 +1,4 @@
-package net.favouriteless.enchanted.util;
+package net.favouriteless.enchanted.common.util;
 
 import com.google.gson.*;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
@@ -17,7 +17,7 @@ public class JsonHelper {
     public static NonNullList<ItemStack> readItemStackList(JsonArray array, boolean readNbt) {
         NonNullList<ItemStack> out = NonNullList.create();
         for (int i = 0; i < array.size(); ++i)
-            out.add(ItemStackHelper.fromJson(array.get(i).getAsJsonObject(), readNbt));
+            out.add(ItemUtils.fromJson(array.get(i).getAsJsonObject(), readNbt));
 
         return out;
     }
@@ -33,7 +33,7 @@ public class JsonHelper {
     public static NonNullList<Item> readItemList(JsonArray array) {
         NonNullList<Item> out = NonNullList.create();
         for (int i = 0; i < array.size(); ++i)
-            out.add(ItemStackHelper.fromJson(array.get(i).getAsJsonObject(), false).getItem());
+            out.add(ItemUtils.fromJson(array.get(i).getAsJsonObject(), false).getItem());
 
         return out;
     }
