+++ common/src/main/java/net/favouriteless/enchanted/common/util/PlayerInventoryHelper.java
@@ -1,4 +1,4 @@
-package net.favouriteless.enchanted.util;
+package net.favouriteless.enchanted.common.util;
 
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
