+++ common/src/main/java/net/favouriteless/enchanted/integrations/jei/EJeiPlugin.java
@@ -10,7 +10,6 @@ import net.favouriteless.enchanted.common.init.registry.EMenuTypes;
 import net.favouriteless.enchanted.common.init.registry.ERecipeTypes;
 import net.favouriteless.enchanted.common.menus.SpinningWheelMenu;
 import net.favouriteless.enchanted.common.menus.WitchOvenMenu;
-import net.favouriteless.enchanted.common.util.RecipeUtils;
 import net.favouriteless.enchanted.integrations.jei.categories.*;
 import net.favouriteless.enchanted.integrations.jei.container_handlers.DistilleryContainerHandler;
 import net.favouriteless.enchanted.integrations.jei.container_handlers.SpinningWheelContainerHandler;
@@ -23,6 +22,7 @@ import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.helpers.IJeiHelpers;
 import mezz.jei.api.registration.*;
+import net.favouriteless.enchanted.common.util.RecipeUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
