+++ common/src/main/java/net/favouriteless/enchanted/integrations/jei/categories/KettleCategory.java
@@ -4,7 +4,7 @@ import net.favouriteless.enchanted.common.Enchanted;
 import net.favouriteless.enchanted.common.init.registry.EItems;
 import net.favouriteless.enchanted.common.recipes.KettleRecipe;
 import net.favouriteless.enchanted.integrations.jei.EJeiRecipeTypes;
-import net.favouriteless.enchanted.util.RecipeUtils;
+import net.favouriteless.enchanted.common.util.RecipeUtils;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
