--- common/src/main/java/net/favouriteless/enchanted/util/RecipeUtils.java
@@ -1,25 +1,0 @@
-package net.favouriteless.enchanted.util;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.multiplayer.ClientLevel;
-import net.minecraft.core.RegistryAccess;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.Recipe;
-
-public class RecipeUtils {
-
-	/**
-	 * Attempts to grab the result of a {@link Recipe} without registry access, using the client's level.
-	 * @return The {@link ItemStack} result of recipe
-	 */
-	public static ItemStack getResultItem(Recipe<?> recipe) {
-		Minecraft minecraft = Minecraft.getInstance();
-		ClientLevel level = minecraft.level;
-		if (level == null) {
-			throw new NullPointerException("level must not be null.");
-		}
-		RegistryAccess registryAccess = level.registryAccess();
-		return recipe.getResultItem(registryAccess);
-	}
-
-}
